+++ simple-config/src/main/java/org/fugerit/java/simple/config/ConfigParams.java
@@ -1,0 +1,11 @@
+package org.fugerit.java.simple.config;
+
+import java.util.Optional;
+
+public interface ConfigParams {
+
+    String getValue( String name );
+
+    Optional<String> getOptionalValue( String name );
+
+}